% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugproductdatabase.R
\name{.dpd_download}
\alias{.dpd_download}
\title{Downloads all sections of the DPD extracts and unzips them}
\usage{
.dpd_download(data_dir)
}
\arguments{
\item{data_dir}{target directory to download to}
}
\value{
none
}
\description{
Downloads all sections of the DPD extracts and unzips them
}
