% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugshortagescanada.R
\name{.dsc_search_all}
\alias{.dsc_search_all}
\title{Return all records from a search}
\usage{
.dsc_search_all(authtoken, query)
}
\arguments{
\item{query}{}
}
\value{

}
\description{
Return all records from a search
}
