% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugshortagescanada.R
\name{.dsc_search_once}
\alias{.dsc_search_once}
\title{A single page of search results}
\usage{
.dsc_search_once(...)
}
\arguments{
\item{...}{name = value pairs of query parameters.}
}
\value{
An R representation of JSON search results via jsonlite::fromJSON
  records.
}
\description{
A single page of search results
}
