% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugshortagescanada.R
\name{.dsc_search_single_page}
\alias{.dsc_search_single_page}
\title{Return a single page of search results}
\usage{
.dsc_search_single_page(...)
}
\arguments{
\item{...}{name = value pairs of query parameters.}
}
\value{
A JSON blog of search results. The 'data' field contains the matching records.
}
\description{
Return a single page of search results
}
