% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugshortagescanada.R
\name{.dsc_tidy_results}
\alias{.dsc_tidy_results}
\title{Tidy up DSC search results}
\usage{
.dsc_tidy_results(results)
}
\arguments{
\item{results}{search results as data.frame}
}
\value{
tibble
}
\description{
Accepts the raw JSON-to-data.frame search results object and converts columns
to their appropriate data types (dates, factors, etc)
}
