% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugproductdatabase.R
\name{dpd_download}
\alias{dpd_download}
\title{Downloads the DPD extracts and unzips them}
\usage{
dpd_download(data_dir = "data/")
}
\arguments{
\item{data_dir}{target directory to download to}
}
\description{
This function downloads all of the available sections (current, inactive,
etc..) of the current DPD extract.
}
