% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugproductdatabase.R
\name{dpd_load}
\alias{dpd_load}
\title{Loads the DPD into memory from a downloaded extract}
\usage{
dpd_load(data_dir = "data/")
}
\arguments{
\item{data_dir}{target directory where extract is stored. Use
dpd_download() to fetch and unpack the latest extract.}
}
\value{
A named list of each table as a data.frame
}
\description{
Each section of the DPD extract is merged together on a per-table basis.
}
