% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugshortagescanada.R
\name{dsc_authtoken}
\alias{dsc_authtoken}
\title{Login to the drugshortagescanada.ca web API}
\usage{
dsc_authtoken(email, password, reuse_authtoken = F)
}
\arguments{
\item{email}{Your email address used to log into the DSC}

\item{password}{Your password}
}
\value{
authentication token
}
\description{
This function logs in to the DSC web API and stores the temporary
authentication token so that it is accessible by dsc_search()
}
\details{
The username and password can be passed into the function, or can be set via
environment variables in .Renviron, or directly like so:

  Sys.setenv('dsc.email' = 'bill.gates@microsoft.com')
  Sys.setenv('dsc.password' = 'passw0rd!')

  dsc_login()
}
