% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugshortagescanada.R
\name{dsc_search}
\alias{dsc_search}
\title{DSC search results}
\usage{
dsc_search(..., format = "tidy")
}
\arguments{
\item{...}{A named list of query parameters.
See the DSC Web API documentation for details on query parameters:
https://www.drugshortagescanada.ca/blog/52}

\item{format}{The default 'tidy' returns a tibble with columns converted to
the appropriate data types. Use 'json' to get results as returned by
the API in JSON format after using jsonlite::rbind_pages to concatenate
multiple pages of results.}
}
\value{
Results formatted according to format
}
\description{
DSC search results
}
