% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugshortagescanada.R
\name{dsc_search_json}
\alias{dsc_search_json}
\title{Return all records from a search as a JSON object}
\usage{
dsc_search_json(authtoken, query)
}
\arguments{
\item{query}{}
}
\value{

}
\description{
Return all records from a search as a JSON object
}
