% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugproductdatabase.R
\name{dpd_load}
\alias{dpd_load}
\title{Loads the DPD}
\usage{
dpd_load(data_dir = "data/", download = T)
}
\arguments{
\item{data_dir}{target directory where extract is stored.}

\item{download}{TRUE if a fresh copy of the extract should be downloaded
before being loaded.}
}
\value{
A named list of each table as a data.frame
}
\description{
Loads the entire Drug Product Database from a cached copy stored on the
filesystem. All tables are returned in a named list, and all sections
(current, inactive, etc) are merged on a per-table bases.
}
