% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugshortagescanada.R
\name{dsc_login}
\alias{dsc_login}
\title{Login}
\usage{
dsc_login(email, password)
}
\arguments{
\item{email}{Your email address used to log into the site}

\item{password}{Your password}
}
\description{
This function logs in to drugshortagescanada.ca via the Web API and returns
an authorization token if successful.
}
\examples{
token = dsc_login("bill.gates@microsoft.com", "passw0rd!")
}
