% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugshortagescanada.R
\name{dsc_search}
\alias{dsc_search}
\title{Return all search results as a flat-ish data.frame}
\usage{
dsc_search(authtoken, query)
}
\arguments{
\item{query}{}
}
\value{

}
\description{
Return all search results as a flat-ish data.frame
}
\examples{

}
